1. Alterados atributos latitude e longitude da classe player para tipo double e adicionados casts (int * 1E6) na instancia do Geopoint 2. Corrigido bug do "Clickable=true" no main.xml

+++ trunk/Cliente/src/br/ufc/model/ClientGameState.java
@@ -18,8 +18,8 @@ public class ClientGameState {
 		// Teste de inicialização dos amigos
 		Player player = new Player("Joao Paulo");
 		
-		player.setLatitude((int)-3730715.0);
-		player.setLongitude((int)-3.8575982E7);
+		player.setLatitude(-3730715.0);
+		player.setLongitude(-3.8575982E7);
 		amigos.add(player);
 	}
 }
